Add UnityMvcApp
This app demonstrates an issue when upgrading an ASP.NET MVC app that uses Unity and Swagger from .NET 6 to .NET 8.  Unity fails to create an EndpointMetadataApiDescriptionProvider because it no longer has a constructor that doesn't have an IServiceProviderIsService required parameter.

The following line can be added to the beginning of Startup.Configure to see whether (and how) the EndpointMetadataApiDescriptionProvider is created.

`IEnumerable<object?> providers = app.ApplicationServices.GetServices(typeof(Microsoft.AspNetCore.Mvc.ApiExplorer.IApiDescriptionProvider));`